miner/requirements.txt
GitPython==3.1.27
numpy==1.22.3
pymongo==4.1.1
requests==2.28.0
